LayoutTests/fast/dom/Element/class-name.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
description('Tests that className works on non HTMLElements');

var htmlElement = document.createElement('div');
htmlElement.setAttribute('class', 'foo');
shouldBeEqualToString('htmlElement.className', 'foo');

var xmlElement = document.createElementNS('http://www.example.com', 'xml');
xmlElement.setAttribute('class', 'foo');
shouldBeEqualToString('xmlElement.className', 'foo');

</script>
<script src="../../js/resources/js-test-post.js"></script>
