third_party/WebKit/LayoutTests/fast/dom/Element/class-name-expected.txt

Tests that className works on non HTMLElements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS htmlElement.className is "foo"
PASS xmlElement.className is "foo"
PASS successfullyParsed is true

TEST COMPLETE
